+++ packages/electrode-archetype-react-app/lib/logger.js
@@ -1,7 +1,9 @@
 "use strict";
 
 const optionalRequire = require("optional-require")(require);
-const devRequire = optionalRequire("electrode-archetype-react-app-dev/require");
+const devRequire =
+  process.env.NODE_ENV !== "production" &&
+  optionalRequire("electrode-archetype-react-app-dev/require");
 
 const winstonLogger = require("./winston-logger");